+++ sdk/spring/spring-messaging-azure-servicebus/src/main/java/com/azure/spring/messaging/servicebus/core/DefaultServiceBusNamespaceProcessorFactory.java
@@ -52,6 +52,7 @@ public final class DefaultServiceBusNamespaceProcessorFactory implements Service
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DefaultServiceBusNamespaceProcessorFactory.class);
     private static final String LOG_IGNORE_NULL_CUSTOMIZER = "The provided customizer is null, will ignore it.";
+    private static final String LOG_IGNORE_NULL_DEDICATED_CUSTOMIZER = "The provided '{}' dedicated customizer is null, will ignore it.";
     private final Map<ConsumerIdentifier, ServiceBusProcessorClient> processorMap = new ConcurrentHashMap<>();
     private final List<Listener> listeners = new ArrayList<>();
     private final NamespaceProperties namespaceProperties;
@@ -277,7 +278,7 @@ public final class DefaultServiceBusNamespaceProcessorFactory implements Service
                                      String subscription,
                                      AzureServiceClientBuilderCustomizer<ServiceBusClientBuilder.ServiceBusProcessorClientBuilder> customizer) {
         if (customizer == null) {
-            LOGGER.debug(LOG_IGNORE_NULL_CUSTOMIZER);
+            LOGGER.debug(LOG_IGNORE_NULL_DEDICATED_CUSTOMIZER, ServiceBusClientBuilder.ServiceBusProcessorClientBuilder.class.getName());
         } else {
             this.dedicatedCustomizers
                 .computeIfAbsent(new ConsumerIdentifier(entityName, subscription), key -> new ArrayList<>())
@@ -297,7 +298,7 @@ public final class DefaultServiceBusNamespaceProcessorFactory implements Service
                                             String subscription,
                                             AzureServiceClientBuilderCustomizer<ServiceBusClientBuilder.ServiceBusSessionProcessorClientBuilder> customizer) {
         if (customizer == null) {
-            LOGGER.debug(LOG_IGNORE_NULL_CUSTOMIZER);
+            LOGGER.debug(LOG_IGNORE_NULL_DEDICATED_CUSTOMIZER, ServiceBusClientBuilder.ServiceBusSessionProcessorClientBuilder.class.getName());
         } else {
             this.dedicatedSessionCustomizers
                 .computeIfAbsent(new ConsumerIdentifier(entityName, subscription), key -> new ArrayList<>())